Drop the whole chain worker state on timeout.
